Added randomisation of game parameters and increasing difficulty

+++ src/main/scala/components/Movement.scala
@@ -1,5 +1,5 @@
 package components
 
-class Movement (var x:Int=0, var y:Int=0) {
+class Movement (var x:Double=0, var y:Double=0) {
 
 }
